Bluetooth: Remove amploader utility, not used
The amploader utility was used to maintain loading and unloading of
the wifi driver to support AMP (BT 3.0+HS) operation on the 1312.
The AMP feature was in the end not supported on this chip and the
functionality of this utility has been disabled for since the feature
was dropped.  For Prima (where AMP feature is supported) a different
mechanism will be used for this functionality, so this utility should
need not be included in the system at all.

Change-Id: Ia0e6755bec6936a34b0753da4d5dc14305cef2a0
CRs-fixed: 336688
(cherry picked from commit bb0b9bd970b7e203213dc1028086dbdc8abac186)